Factor out objective functions from FitUtil
Move each evaluation class in its each own file, rename Eval class.
