+++ math/mathcore/inc/Fit/LogLikelihoodFCN.h
@@ -19,7 +19,7 @@
 
 #include "Fit/UnBinData.h"
 
-#include "Fit/FitUtil.h"
+#include "Fit/EvaluateLogL.hxx"
 
 #include <memory>
 
@@ -126,8 +126,8 @@ public:
    // need to be virtual to be instantited
    virtual void Gradient(const double *x, double *g) const {
       // evaluate the chi2 gradient
-      FitUtil::Evaluate<typename BaseFCN::T>::EvalLogLGradient(BaseFCN::ModelFunction(), BaseFCN::Data(), x, g,
-                                                               fNEffPoints, fExecutionPolicy);
+      FitUtil::LogL<typename BaseFCN::T>::EvalGradient(BaseFCN::ModelFunction(), BaseFCN::Data(), x, g, fNEffPoints,
+                                                       fExecutionPolicy);
    }
 
    /// get type of fit method function
@@ -154,7 +154,8 @@ private:
     */
    virtual double DoEval (const double * x) const {
       this->UpdateNCalls();
-      return FitUtil::Evaluate<T>::EvalLogL(BaseFCN::ModelFunction(), BaseFCN::Data(), x, fWeight, fIsExtended, fNEffPoints, fExecutionPolicy);
+      return FitUtil::LogL<T>::Eval(BaseFCN::ModelFunction(), BaseFCN::Data(), x, fWeight, fIsExtended, fNEffPoints,
+                                    fExecutionPolicy);
    }
 
    // for derivatives
